requirements.txt
requests
rich
pillow
setuptools

